Split huge perform_logic method into own module
