+++ lib/src/story/mod.rs
@@ -222,6 +222,7 @@ mod misc {
 }
 
 mod choices;
+mod control_logic;
 pub mod errors;
 pub mod external_functions;
 mod flow;
